refactor store code
