+++ api/errors.go
@@ -2,6 +2,8 @@ package api
 
 import (
 	"encoding/json"
+	"fmt"
+	"github.com/go-playground/validator"
 	"net/http"
 )
 
@@ -25,4 +27,14 @@ func writeError(w http.ResponseWriter, errs jsonErrors, httpCode int) {
 
 func buildErrors(err error, code string) jsonErrors {
 	return jsonErrors{Error: []jsonError{{Description: err.Error(), Code: code}}}
+}
+
+func buildValidationErrors(validationErrs validator.ValidationErrors) jsonErrors {
+	jsonErrs := jsonErrors{Error: []jsonError{}}
+
+	for _, validationErr := range validationErrs {
+		desc := fmt.Sprintf("%s %s", validationErr.Field(), validationErr.Tag())
+		jsonErrs.Error = append(jsonErrs.Error, jsonError{Code: "ValidationError", Description: desc})
+	}
+	return jsonErrs
 }
